Add an integer range analysis
The integer range analysis relies on traits being defined
on operations to compute the range of results and basic block
arguments. For now, only non-region operations are defined.

This PR also includes a test pass to dump the results of the
integer range analysis, which we use in our test.